requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
cryptography>=41.0.0
pywin32>=307
psutil>=5.9.0
schedule>=1.2.0
jinja2>=3.1.0
werkzeug>=2.3.0
qrcode>=7.4.0
Pillow>=9.0.0
pandas>=1.5.0
openpyxl>=3.0.0
